Add blog section (#40)
* feat(section): add blog section design

Adding a blog post with a very long title would cause the card layout
to increase it's size, making it inconsistent with the rest of the
cards. So I added a new CSS class and new tailwind classes to wrap the
title and also long tags. It was also added a `title` attribute so when
hovering with a mouse it's possible to read the longer text version.

Since the `BlogPosts` component is going to be an optional feature, I
added a new boolean prop called `enabled`, if this is false, the
component won't be rendered.

The props that the Button component receives are not optional and cannot
be null or undefined, with this patch I added nullish coalescing
operator and also a ternary to avoid the component crashing if it does
not receive a style or an icon. Affected file: